backend/requirements.txt
Django>=4.2,<5.0
djangorestframework-simplejwt
django-cors-headers
psycopg[binary]
stripe
django-environ

# Dev / Testing
pytest
pytest-django
black
flake8
mypy
python-dotenv

